+++ app/javascript/vue/tasks/loans/new/store/actions/loadLoan.js
@@ -7,6 +7,6 @@ export default ({ commit, state, dispatch }, id) => {
   return Loan.find(id, { extend: ['roles'] }).then((response) => {
     commit(MutationNames.SetLoading, false)
     commit(MutationNames.SetLoan, response.body)
-    dispatch(ActionNames.LoadLoanItems, id)
+    dispatch(ActionNames.LoadLoanItems, { loanId: id })
   })
 }