+++ app/javascript/vue/tasks/loans/new/store/getters/getters.js
@@ -2,22 +2,22 @@ import getLoan from './getLoan'
 import getLoanItems from './getLoanItems'
 import getEditLoanItems from './getEditLoanItems'
 import getSettings from './getSettings'
+import getPagination from './getPagination'
 
 const GetterNames = {
   GetLoan: 'getLoan',
   GetLoanItems: 'getLoanItems',
   GetEditLoanItems: 'getEditLoanItems',
-  GetSettings: 'getSettings'
+  GetSettings: 'getSettings',
+  GetPagination: 'getPagination'
 }
 
 const GetterFunctions = {
   [GetterNames.GetLoan]: getLoan,
   [GetterNames.GetLoanItems]: getLoanItems,
   [GetterNames.GetEditLoanItems]: getEditLoanItems,
-  [GetterNames.GetSettings]: getSettings
+  [GetterNames.GetSettings]: getSettings,
+  [GetterNames.GetPagination]: getPagination
 }
 
-export {
-  GetterNames,
-  GetterFunctions
-}
+export { GetterNames, GetterFunctions }